#a5 2 wolkenkrabber opmaak+delen behalen

+++ app/src/main/java/com/example/faithandroid/skyscraper/GoalDetailsFragment.kt
@@ -1,11 +1,15 @@
 package com.example.faithandroid.skyscraper
 
+import android.graphics.Color
+import android.os.Build
 import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
+import android.widget.Toast
+import androidx.annotation.RequiresApi
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.DialogFragment
 import androidx.lifecycle.Observer
@@ -15,8 +19,14 @@ import androidx.navigation.fragment.navArgs
 import com.example.faithandroid.skyscraper.GoalDetailsFragmentArgs
 import com.example.faithandroid.R
 import com.example.faithandroid.databinding.SkyscraperGoaldetailsBinding
+import com.google.android.material.card.MaterialCardView
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.android.material.snackbar.Snackbar
 import kotlinx.android.synthetic.main.skyscraper_goalpostimage.view.*
+import java.time.LocalDate
+import java.time.LocalDateTime
+import java.time.format.DateTimeFormatter
+import java.time.format.FormatStyle
 
 class GoalDetailsFragment: DialogFragment() {
     val args: GoalDetailsFragmentArgs by navArgs()
@@ -26,6 +36,7 @@ class GoalDetailsFragment: DialogFragment() {
         super.onCreate(savedInstanceState)
     }
 
+    @RequiresApi(Build.VERSION_CODES.O)
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
@@ -45,30 +56,66 @@ class GoalDetailsFragment: DialogFragment() {
         //val faithProperty = goalDetailsFragmentArgs.fromBundle(requireArguments()).title
 
 
-
         binding.titelText.text = args.goal.title
         binding.beschrijvingText.text = args.goal.description
+        binding.goalShared = args.goal.shared
+        binding.goalDetail = args.goal.completed
+
+        var localdate = LocalDateTime.parse(args.goal.date)
+        var date = localdate.dayOfMonth.toString() +" "+ localdate.month.toString()+ " "+ localdate.year.toString()
+        binding.datumText.text = date
 
         binding.btnBehaald.setOnClickListener { view: View ->
             viewModel.goalBehaald(args.goal.id)
+            view.findNavController().navigate(R.id.skyscraperFragment)
         }
 
         binding.btnDelen.setOnClickListener { view: View ->
             viewModel.shareGoal(args.goal.id)
+           view.findNavController().navigate(R.id.billboardFragment)
         }
 
         binding.btnVerwijder.setOnClickListener { view: View ->
-            viewModel.deleteGoal(args.goal.id)
+          if (args.goal.shared) {
+
+                MaterialAlertDialogBuilder(view.getContext())
+                    .setTitle("Goal verwijderen")
+                    .setMessage("Deze goal is gedeeld. Ben je zeker dat je deze goal wilt verwijderen? ")
+                    .setPositiveButton("Ja") { _, which ->
+                        // unshare goal en then delete goal
+                        viewModel.shareGoal(args.goal.id)
+                        viewModel.deleteGoal(args.goal.id)
+                        view.findNavController().navigate(R.id.skyscraperFragment)
+                    }
+                    .setNegativeButton("Nee") { _, which ->
+                        // nothing has to happen here
+                    }.show()
+            }else if(args.goal.completed){
+              Snackbar.make(view, "Deze goal is behaald, u kan deze dus niet verwijderen.", Snackbar.LENGTH_SHORT).setAction(
+               "" )
+              {
+              }.show()
+        }
+          else {
+                viewModel.deleteGoal(args.goal.id)
+                view.findNavController().navigate(R.id.skyscraperFragment)
+            }
         }
 
         viewModel.shareStatus.observe(this.viewLifecycleOwner, Observer {
             val contextView = this.view
             if (contextView != null) {
-                Snackbar.make(contextView, viewModel.shareStatus.value.toString(), Snackbar.LENGTH_SHORT).setAction(
-                    R.string.tryAgain
+                Toast.makeText(
+                    context,
+                    "Je doel is gedeeld",
+                    Toast.LENGTH_LONG
+                ).show()
+             }
+            else {
+                Snackbar.make(contextView!!, viewModel.shareStatus.value.toString(), Snackbar.LENGTH_SHORT).setAction(
+                    "Probeer opnieuw"
                 )
                 {
-                    viewModel.shareGoal(args.goal.id)
                 }.show()
             }
         })
@@ -76,11 +123,16 @@ class GoalDetailsFragment: DialogFragment() {
         viewModel.completedStatus.observe(this.viewLifecycleOwner, Observer {
             val contextView = this.view
             if (contextView != null) {
-                Snackbar.make(contextView, viewModel.completedStatus.value.toString(), Snackbar.LENGTH_SHORT).setAction(
-                    R.string.tryAgain
+                Toast.makeText(
+                    context,
+                    "Je doel is behaald",
+                    Toast.LENGTH_LONG
+                ).show()
+            } else {
+                Snackbar.make(contextView!!, viewModel.completedStatus.value.toString(), Snackbar.LENGTH_SHORT).setAction(
+                    "Probeer opnieuw"
                 )
                 {
-                    viewModel.goalBehaald(args.goal.id)
                 }.show()
             }
         })
@@ -88,13 +140,20 @@ class GoalDetailsFragment: DialogFragment() {
         viewModel.removeStatus.observe(this.viewLifecycleOwner, Observer {
             val contextView = this.view
             if (contextView != null) {
-                Snackbar.make(contextView, viewModel.removeStatus.value.toString(), Snackbar.LENGTH_SHORT).setAction(
-                    R.string.tryAgain
+                Toast.makeText(
+                    context,
+                    "Je doel is verwijderd",
+                    Toast.LENGTH_LONG
+                ).show()
+            }
+            else {
+                Snackbar.make(contextView!!, viewModel.removeStatus.value.toString(), Snackbar.LENGTH_SHORT).setAction(
+                    "Probeer opnieuw"
                 )
                 {
-                    viewModel.deleteGoal(args.goal.id)
                 }.show()
             }
+
         })
 
         args.goal.steps.forEach{step ->
@@ -117,5 +176,5 @@ class GoalDetailsFragment: DialogFragment() {
         return binding.root
         * */
 
-
 }
+
